Add runners abstract factory
The runners abstract factory allows to use factories
for all formats supported by /api/runs. The only
implemented format is "Pipeline", but it allows to
support "Link", which allows to add a run on an existing
pipeline.

This may be an over-engineered design for now, but it
allowed me to practice the abstract factory pattern in
go.