Simplify a lot the pipeline resolution in run
The runner factory would have been suitable if there was
something else than a pipeline that could be run.
It was designed to resolve pipeline links, but this can
be done by the pipeline package itself, when creating
a pipeline from a spec. If the spec is a link, then the
pipeline factory resolves the link, and returns a pipeline.

This simplification also removes the concept of runner,
which gives the run responsibility to the pipeline itself.