Add status to runs
The runs were replaced in redis from string to hash,
containing the following fields:
- uid
- status

Having a status in runs allows to set the run status
as "canceled", and to avoid deducing information that
can not always been deduced from jobs status.

Change the work queue from jobs:work to runs:work,
because workers now consume runs instead of jobs,
which allows in case of cancellation to have a unity
for a single run.