Update the dependencies storage in redis
As dependencies are stored in an array, it is possible
to have multiple dependencies to the same job.
This was fixed by changing the dependencies format from
`dependency:<dep>:job:<name>:run:<uid>` to
`dependency:<index>:job:<name>:run:<uid>`.
Dependencies now also contain a field for the dependency
job key.