Add k8s cloud provider
The k8s cloud provider allows to start and monitor jobs
on Kubernetes.
It is initialized from the in-cluster configuration,
or the KUBECONFIG environment variable if set.
After a job is completed, it is deleted, along with all
associated pods.
It could not be properly tested, because most of it is
calls to the Kubernetes API, and it does not provide
logic.